flisol_inscricoes/requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
django-discover-runner==1.0

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.1
